# fct-frontend/api-spec/openapi.yaml
openapi: 3.1.1
info:
    title: REST API for the website of "Academia Fabiana Perez"
    description: REST API for the website of "Academia Fabiana Perez"
    version: 1.0.0
servers:
    - url: http://localhost:8080/api/v1
paths:
...
    /courses:
        get:
            summary: Get all courses
            operationId: getAllCourses
            tags:
                - Courses
            parameters:
                - name: userId
                  in: query
                  description: Filter by user id
                  required: false
                  schema:
                      type: integer
                      format: int64
                - name: categoryId
                  in: query
                  description: Filter by category id
                  required: false
                  schema:
                      type: integer
                      format: int64
            responses:
                200:
                    description: Successful operation
                    content:
                        application/json:
                            schema:
                                type: array
                                items:
                                    $ref: "#/components/schemas/Course"
        post:
            summary: Add a new course
            operationId: addCourse
            tags:
                - Courses
            requestBody:
                description: Course to add
                required: true
                content:
                    application/json:
                        schema:
                            $ref: "#/components/schemas/AddCourseRequest"
            responses:
                201:
                    description: Successful operation
                    content:
                        application/json:
                            schema:
                                $ref: "#/components/schemas/Course"
                400:
                    description: Bad request
                    content:
                        application/json:
                            schema:
                                $ref: "#/components/schemas/ValidationErrorMessage"
                401:
                    description: Unauthorized
                    content:
                        application/json:
                            schema:
                                $ref: "#/components/schemas/ErrorMessage"
                403:
                    description: Forbidden
                    content:
                        application/json:
                            schema:
                                $ref: "#/components/schemas/ErrorMessage"
                409:
                    description: Course already exists
                    content:
                        application/json:
                            schema:
                                $ref: "#/components/schemas/ErrorMessage"
    /courses/{id}:
        get:
            summary: Get a course by id
            operationId: getCourseById
            tags:
                - Courses
            parameters:
                - name: id
                  in: path
                  description: Course id
                  required: true
                  schema:
                      type: integer
                      format: int64
            responses:
                200:
                    description: Successful operation
                    content:
                        application/json:
                            schema:
                                $ref: "#/components/schemas/Course"
                400:
                    description: Bad request
                    content:
                        application/json:
                            schema:
                                $ref: "#/components/schemas/ErrorMessage"
                401:
                    description: Unauthorized
                    content:
                        application/json:
                            schema:
                                $ref: "#/components/schemas/ErrorMessage"
                403:
                    description: Forbidden
                    content:
                        application/json:
                            schema:
                                $ref: "#/components/schemas/ErrorMessage"
                404:
                    description: Course not found
                    content:
                        application/json:
                            schema:
                                $ref: "#/components/schemas/ErrorMessage"
        put:
            summary: Update a course
            operationId: updateCourse
            tags:
                - Courses
            parameters:
                - name: id
                  in: path
                  description: Course id
                  required: true
                  schema:
                      type: integer
                      format: int64
            requestBody:
                description: Course to update
                required: true
                content:
                    application/json:
                        schema:
                            $ref: "#/components/schemas/UpdateCourseRequest"
            responses:
                200:
                    description: Successful operation
                    content:
                        application/json:
                            schema:
                                $ref: "#/components/schemas/Course"
                400:
                    description: Bad request
                    content:
                        application/json:
                            schema:
                                $ref: "#/components/schemas/ValidationErrorMessage"
                401:
                    description: Unauthorized
                    content:
                        application/json:
                            schema:
                                $ref: "#/components/schemas/ErrorMessage"
                403:
                    description: Forbidden
                    content:
                        application/json:
                            schema:
                                $ref: "#/components/schemas/ErrorMessage"
                404:
                    description: Course not found
                    content:
                        application/json:
                            schema:
                                $ref: "#/components/schemas/ErrorMessage"
        delete:
            summary: Delete a course
            operationId: deleteCourse
            tags:
                - Courses
            parameters:
                - name: id
                  in: path
                  description: Course id
                  required: true
                  schema:
                      type: integer
                      format: int64
            responses:
                204:
                    description: Successful operation
                400:
                    description: Bad request
                    content:
                        application/json:
                            schema:
                                $ref: "#/components/schemas/ErrorMessage"
                401:
                    description: Unauthorized
                    content:
                        application/json:
                            schema:
                                $ref: "#/components/schemas/ErrorMessage"
                403:
                    description: Forbidden
                    content:
                        application/json:
                            schema:
                                $ref: "#/components/schemas/ErrorMessage"
                404:
                    description: Course not found
                    content:
                        application/json:
                            schema:
                                $ref: "#/components/schemas/ErrorMessage"
    /courses/img/{id}:
        put:
            summary: Update course image
            operationId: updateCourseImage
            tags:
                - Courses
            parameters:
                - name: id
                  in: path
                  description: Course id
                  required: true
                  schema:
                      type: integer
                      format: int64
            requestBody:
                description: Course image to update
                required: true
                content:
                    multipart/form-data:
                        schema:
                            type: object
                            properties:
                                img:
                                    type: string
                                    format: binary
                            required: [img]
            responses:
                200:
                    description: Successful operation
                    content:
                        application/json:
                            schema:
                                $ref: "#/components/schemas/Course"
                400:
                    description: Bad request
                    content:
                        application/json:
                            schema:
                                $ref: "#/components/schemas/ValidationErrorMessage"
                401:
                    description: Unauthorized
                    content:
                        application/json:
                            schema:
                                $ref: "#/components/schemas/ErrorMessage"
                403:
                    description: Forbidden
                    content:
                        application/json:
                            schema:
                                $ref: "#/components/schemas/ErrorMessage"
                404:
                    description: Course not found
                    content:
                        application/json:
                            schema:
                                $ref: "#/components/schemas/ErrorMessage"
        delete:
            summary: Delete course image
            operationId: removeCourseImage
            tags:
                - Courses
            parameters:
                - name: id
                  in: path
                  description: Course id
                  required: true
                  schema:
                      type: integer
                      format: int64
                - name: imgName
                  in: query
                  description: Image name
                  required: true
                  schema:
                      type: string
            responses:
                200:
                    description: Successful operation
                400:
                    description: Bad request
                    content:
                        application/json:
                            schema:
                                $ref: "#/components/schemas/ValidationErrorMessage"
                401:
                    description: Unauthorized
                    content:
                        application/json:
                            schema:
                                $ref: "#/components/schemas/ErrorMessage"
                403:
                    description: Forbidden
                    content:
                        application/json:
                            schema:
                                $ref: "#/components/schemas/ErrorMessage"
                404:
                    description: Course not found
                    content:
                        application/json:
                            schema:
                                $ref: "#/components/schemas/ErrorMessage"
    /courses/{courseId}/user/{userId}:
        post:
            summary: Add user to course
            operationId: addUserToCourse
            tags:
                - Courses
            parameters:
                - name: courseId
                  in: path
                  description: The unique identifier of the course to add the user to.
                  required: true
                  schema:
                      type: integer
                      format: int64
                - name: userId
                  in: path
                  description: The unique identifier of the user to add to the course.
                  required: true
                  schema:
                      type: integer
                      format: int64
            responses:
                200:
                    description: Successful operation
                    content:
                        application/json:
                            schema:
                                $ref: "#/components/schemas/CourseUser"
                400:
                    description: Invalid input
                    content:
                        application/json:
                            schema:
                                $ref: "#/components/schemas/ErrorMessage"
                401:
                    description: Unauthorized
                    content:
                        application/json:
                            schema:
                                $ref: "#/components/schemas/ErrorMessage"
                403:
                    description: Forbidden
                    content:
                        application/json:
                            schema:
                                $ref: "#/components/schemas/ErrorMessage"
                404:
                    description: Course or user not found
                    content:
                        application/json:
                            schema:
                                $ref: "#/components/schemas/ErrorMessage"
        patch:
            summary: Update user status on course
            operationId: updateUserStatusOnCourse
            tags:
                - Courses
            parameters:
                - name: courseId
                  in: path
                  description: Course id
                  required: true
                  schema:
                      type: integer
                      format: int64
                - name: userId
                  in: path
                  description: User id
                  required: true
                  schema:
                      type: integer
                      format: int64
            requestBody:
                description: User status
                required: true
                content:
                    application/json:
                        schema:
                            $ref: "#/components/schemas/UpdateUserStatusOnCourseRequest"
            responses:
                200:
                    description: Successful operation
                    content:
                        application/json:
                            schema:
                                $ref: "#/components/schemas/CourseUser"
                400:
                    description: Bad request
                    content:
                        application/json:
                            schema:
                                $ref: "#/components/schemas/ErrorMessage"
                401:
                    description: Unauthorized
                    content:
                        application/json:
                            schema:
                                $ref: "#/components/schemas/ErrorMessage"
                404:
                    description: Course or user not found
                    content:
                        application/json:
                            schema:
                                $ref: "#/components/schemas/ErrorMessage"
...
components:
    schemas:
...
        UserCourseEnrollmentStatusEnum:
            type: string
            enum: [WAITING, ENROLLED]
            x-field-extra-annotation: "@ValidEnum(UserCourseEnrollmentStatusEnum.class)"

        # ---- Objects ----
...
        Course:
            type: object
            properties:
                id:
                    type: integer
                    format: int64
                startDate:
                    type: string
                    format: date
                endDate:
                    type: string
                    format: date
                enrollmentPrice:
                    type: number
                    format: double
                categoryId:
                    type: integer
                    format: int64
                description:
                    type: string
                imgNames:
                    type: array
                    items:
                        type: string
            required:
                [
                    id,
                    startDate,
                    endDate,
                    enrollmentPrice,
                    categoryId,
                    description,
                    imgNames,
                ]
...
        CourseUser:
            type: object
            properties:
                userId:
                    type: integer
                    format: int64
                courseId:
                    type: integer
                    format: int64
                status:
                    $ref: "#/components/schemas/UserCourseEnrollmentStatusEnum"
            required: [userId, courseId, status]
...
        AddCourseRequest:
            type: object
            properties:
                startDate:
                    type: string
                    format: date
                    x-field-extra-annotation: >-
                        @Future(message = "Start date must be in the future")
                        @DateTimeFormat(iso = DateTimeFormat.ISO.DATE_TIME)
                endDate:
                    type: string
                    format: date
                    x-field-extra-annotation: >-
                        @Future(message = "End date must be in the future")
                        @DateTimeFormat(iso = DateTimeFormat.ISO.DATE_TIME)
                enrollmentPrice:
                    type: number
                    format: double
                    minimum: 1
                categoryId:
                    type: integer
                    format: int64
                description:
                    x-field-extra-annotation: '@NotBlank(message = "Description field cannot be empty")'
                    type: string
            required:
                [startDate, endDate, enrollmentPrice, categoryId, description]
        UpdateCourseRequest:
            type: object
            properties:
                startDate:
                    type: string
                    format: date
                    x-field-extra-annotation: >-
                        @Future(message = "Start date must be in the future")
                        @DateTimeFormat(iso = DateTimeFormat.ISO.DATE_TIME)
                endDate:
                    type: string
                    format: date
                    x-field-extra-annotation: >-
                        @Future(message = "End date must be in the future")
                        @DateTimeFormat(iso = DateTimeFormat.ISO.DATE_TIME)
                enrollmentPrice:
                    type: number
                    format: double
                    minimum: 1
                categoryId:
                    type: integer
                    format: int64
                    minimum: 1
                description:
                    x-field-extra-annotation: '@NotEmpty(message = "Description field cannot be empty")'
                    type: string
        UpdateUserStatusOnCourseRequest:
            type: object
            properties:
                status:
                    $ref: "#/components/schemas/UserCourseEnrollmentStatusEnum"
            required: [status]

        ## ---- CourseCategory ----
...
        ErrorMessage:
            type: object
            properties:
                timestamp:
                    type: string
                    format: date-time
                    example: "1970-01-01T00:00+18:00"
                status:
                    type: integer
                    format: int32
                    example: 404
                error:
                    type: string
                    example: "Not Found"
                message:
                    type: string
                    example: "Resource not found"
                path:
                    type: string
                    example: "/api/v1/resource/1"
            required: [timestamp, status, error, message, path]
        ValidationErrorMessage:
            type: object
            properties:
                timestamp:
                    type: string
                    format: date-time
                    example: "1970-01-01T00:00+18:00"
                status:
                    type: integer
                    format: int32
                    example: 404
                error:
                    type: string
                    example: "Bad Request"
                messages:
                    type: object
                    additionalProperties:
                        type: string
                    example:
                        "name": "Name field cannot be blank"
                path:
                    type: string
                    example: "/api/v1/resource/1"
            required: [timestamp, status, error, messages, path]
    securitySchemes:
        BearerAuth:
            description: Bearer token using a JWT
            type: http
            scheme: bearer
            bearerFormat: JWT